backend/requirements.txt
Django==5.2
strawberry-graphql==0.275.1
strawberry-graphql-Django==0.62.0
graphql-core==3.2.6
django-cors-headers==4.7.0
cryptography
psycopg2-binary
django-allauth
requests
google-auth==2.23.4
PyJWT==2.8.0
dotenv
django-taggit==6.1.0
django-storages
boto3
python-multipart